+++ proxy/src/redis.rs
@@ -1,4 +1,4 @@
+pub mod cancellation_publisher;
 pub mod connection_with_credentials_provider;
 pub mod elasticache;
 pub mod notifications;
-pub mod publisher;
